+++ src/parsers.js
@@ -267,11 +267,11 @@ const parseDomainWhois = (domain, whois) => {
 
 	// Parse WHOIS info for specific TLDs
 
-	if (domain.endsWith('.uk') || domain.endsWith('.be') || domain.endsWith('.nl') || domain.endsWith('.eu') || domain.endsWith('.ly') || domain.endsWith('.mx')|| domain.endsWith('.gg')) {
+	if (domain.endsWith('.uk') || domain.endsWith('.be') || domain.endsWith('.nl') || domain.endsWith('.eu') || domain.endsWith('.ly') || domain.endsWith('.mx') || domain.endsWith('.gg') || domain.endsWith('.je') || domain.endsWith('.as')) {
 		lines = handleMultiLines(lines)
 	}
 
-	if (domain.endsWith('.gg')) {
+	if (domain.endsWith('.gg') || domain.endsWith('.je') || domain.endsWith('.as')) {
 		lines = handleMissingColons(lines)
 	}